- project cleanup and removed unused texture asset

+++ dreco-engine/src/renderer/opengles2_renderer.hxx
@@ -23,7 +23,7 @@ public:
 
 	void UpdateViewportSize();
 
-	int GetStencilIndexFromPixel(const vec2& _p_coord);
+	uint8_t GetStencilIndexFromPixel(const vec2& _p_coord);
 
 	void DrawScene();
